+++ modules/genealogy-dna-livewire/src/DnaLivewireServiceProvider.php
@@ -17,9 +17,13 @@ final class DnaLivewireServiceProvider extends ServiceProvider
         Livewire::component('module-genealogy-dna::match-list', DnaMatchList::class);
         Livewire::component('module-genealogy-dna::group-list', DnaGroupList::class);
         Livewire::component('module-genealogy-dna::annotation-list', DnaAnnotationList::class);
+        Livewire::component('module-genealogy-dna::segment-list', DnaSegmentList::class);
+        Livewire::component('module-genealogy-dna::consent-list', DnaConsentList::class);
         Livewire::component('genealogy-dna-match-list', DnaMatchList::class);
         Livewire::component('genealogy-dna-group-list', DnaGroupList::class);
         Livewire::component('genealogy-dna-annotation-list', DnaAnnotationList::class);
+        Livewire::component('genealogy-dna-segment-list', DnaSegmentList::class);
+        Livewire::component('genealogy-dna-consent-list', DnaConsentList::class);
     }
 }